Fix log handling of php image so logs are sent to docker
Fix the log handling of the php base image so logs are correctly sent to
docker. This is done by making fpm capture the output from each php
worker and sending this to /dev/stderr. The entrypoint script must also
run exec to replace itself with fpm to correctly inherit the stderr
device linked to docker for process 1.