Add PHP grpc extension and grpc watcher image
Add grpc and protobuf PHP extension to the PHP base image for
communicating with gRPC servers in PHP projects.

Add a grpc watcher/compiler image to compile PHP client side code from
proto files. Allow PHP namespaces to be dropped from the generated
directory structure by specifying a root path that is copied in. This
is for the benefit of Laravel, which uses an App top level
namespace in a folder called app.