Add new combined deskpro base Docker image
Add new combined Deskkpro base Docker image which takes php and fpm and
combines in to a single Docker image using supervisord to manage the two
processes.

Add an event listener script which supervisord will trigger in the event
of an unexpected process exit which can terminate the supervisord
process and cause a running docker container to exit.

Setup the log shipping for nginx and fpm processes to send their logs to
docker log subsystem.

Have fpm listen on a socket instead of a tcp network port

Allow the user used to run the fpm process in the base-php image to be
overridden using environment variables. This allows the passwd/group
files to be bind mounted to the host and another user to be given.

This is most useful during development in Vagrant, where the container
is made to run as the vagrant user, giving r/w perms to access all files
and improving DX.