Allow fpm user override
Allow the user used to run the fpm process in the base-php image to be
overridden using environment variables. This allows the passwd/group
files to be bind mounted to the host and another user to be given.

This is most useful during development in Vagrant, where the container
is made to run as the vagrant user, giving r/w perms to access all files
and improving DX.